+++ src/main/kotlin/com/aldikitta/routes/PostRoutes.kt
@@ -27,28 +27,23 @@ fun Route.createPost(
                     call.respond(HttpStatusCode.BadRequest)
                     return@post
                 }
-
-                ifEmailBelongsToUser(
-                    userId = request.userId,
-                    validateEmail = userService::doesEmailBelongToUserId
-                ) {
-                    val didUserExist = postService.createPostIfUserExists(request)
-                    if (!didUserExist) {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = false,
-                                message = USER_NOT_FOUND
-                            )
+                val userId = call.userId
+                val didUserExist = postService.createPostIfUserExists(request, userId)
+                if (!didUserExist) {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = USER_NOT_FOUND
                         )
-                    } else {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BasicApiResponse(
-                                successful = true,
-                            )
+                    )
+                } else {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true,
                         )
-                    }
+                    )
                 }
             }
         }
@@ -57,29 +52,19 @@ fun Route.createPost(
 
 fun Route.getPostsForFollows(
     postService: PostService,
-    userService: UserService,
 ) {
     authenticate {
         route("/api/post/get") {
             get {
-                val userId = call.parameters[QueryParams.PARAM_USER_ID] ?: kotlin.run {
-                    call.respond(HttpStatusCode.BadRequest)
-                    return@get
-                }
                 val page = call.parameters[QueryParams.PARAM_PAGE]?.toIntOrNull() ?: 0
                 val pageSize =
                     call.parameters[QueryParams.PARAM_PAGE_SIZE]?.toIntOrNull() ?: Constants.DEFAULT_POST_PAGE_SIZE
 
-                ifEmailBelongsToUser(
-                    userId = userId,
-                    validateEmail = userService::doesEmailBelongToUserId
-                ) {
-                    val posts = postService.getPostsForFollows(userId = userId, page = page, pageSize = pageSize)
-                    call.respond(
-                        HttpStatusCode.OK,
-                        posts
-                    )
-                }
+                val posts = postService.getPostsForFollows(userId = call.userId, page = page, pageSize = pageSize)
+                call.respond(
+                    HttpStatusCode.OK,
+                    posts
+                )
             }
         }
     }
@@ -87,32 +72,35 @@ fun Route.getPostsForFollows(
 
 fun Route.deletePost(
     postService: PostService,
-    userService: UserService,
     likeService: LikeService
 ) {
-    route("/api/post/delete") {
-        delete {
-            val request = call.receiveNullable<DeletePostRequest>() ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
-                return@delete
-            }
+    authenticate {
+        route("/api/post/delete") {
+            delete {
+                val request = call.receiveNullable<DeletePostRequest>() ?: kotlin.run {
+                    call.respond(HttpStatusCode.BadRequest)
+                    return@delete
+                }
 
-            val post = postService.getPost(request.postId)
-            if (post == null) {
-                call.respond(
-                    HttpStatusCode.NotFound
-                )
-                return@delete
-            }
-            ifEmailBelongsToUser(
-                userId = post.userId,
-                validateEmail = userService::doesEmailBelongToUserId
-            ) {
-                postService.deletePost(request.postId)
-                likeService.deleteLikesForParent(request.postId)
+                val post = postService.getPost(request.postId)
+                if (post == null) {
+                    call.respond(
+                        HttpStatusCode.NotFound
+                    )
+                    return@delete
+                }
+                if (post.userId == call.userId) {
+                    postService.deletePost(request.postId)
+                    likeService.deleteLikesForParent(request.postId)
 
-                // TODO: Delete comments from post
-                call.respond(HttpStatusCode.OK)
+                    // TODO: Delete comments from post
+                    call.respond(HttpStatusCode.OK)
+                } else {
+                    call.respond(
+                        HttpStatusCode.Unauthorized
+                    )
+                }
+
             }
         }
     }
